% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aureleutils-package.R
\docType{package}
\name{aureleutils-package}
\alias{aureleutils}
\alias{aureleutils-package}
\title{aureleutils: Personal Utility Functions}
\description{
A collection of various utility functions.
}
\author{
\strong{Maintainer}: Aurèle \email{first.last@example.com}

}
\keyword{internal}
