% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install, don't update}
\usage{
install(pkg)
}
\arguments{
\item{pkg}{Package name, character}
}
\description{
Install, don't update
}
\examples{
\dontrun{
install("remotes")
}

}
